prince kumar
This program takes an arithmetic operator (+, -, *, /) and two operands as the input. It then checks whether the user wants to quit the program if yes then quit it, for this, we can use a special character and tell the user about it, like here we used “x”. The operation is performed based on the operator.